dynamic colors -- update colors every time a window name changes
Attempt to update colors for everything using a "wincolorlist",
including titles, icon titles, icon manager entries, and workspace
mini-windows, every time they are drawn so as to take into account any
possible changes in the window name.

Code for updating window title colors was already present but commented
out because it was missing some initialization bits and it was in the
wrong place.